+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `html` option for tooltips
 - Added `getMapObject` and `destroy` functions
 - Added `maps` property
+- Added `accessToken` option
 - Updated default style to `streets-v12`
 - Reduced dependencies
 - Dropped support for AJAX through jQuery and Zepto